+++ xpaw/utils.py
@@ -1,0 +1,119 @@
+# coding=utf-8
+
+import os
+import re
+import cgi
+import hashlib
+import logging
+from importlib import import_module
+from pkgutil import iter_modules
+from os.path import isfile, join, abspath
+import string
+
+from xpaw.config import Config
+from xpaw.cluster import LocalCluster
+
+
+def load_object(path):
+    dot = path.rindex(".")
+    module, name = path[:dot], path[dot + 1:]
+    mod = import_module(module)
+    return getattr(mod, name)
+
+
+def walk_modules(path):
+    mods = []
+    mod = import_module(path)
+    mods.append(mod)
+    if hasattr(mod, "__path__"):
+        for _, subpath, ispkg in iter_modules(mod.__path__):
+            fullpath = path + "." + subpath
+            if ispkg:
+                mods += walk_modules(fullpath)
+            else:
+                submod = import_module(fullpath)
+                mods.append(submod)
+    return mods
+
+
+def configure_logging(name, config):
+    logger = logging.getLogger(name)
+    logger.setLevel(config["log_level"])
+    log_stream_handler = logging.StreamHandler()
+    log_stream_handler.setLevel(config["log_level"])
+    log_formatter = logging.Formatter(config["log_format"], config["log_dateformat"])
+    log_stream_handler.setFormatter(log_formatter)
+    logger.addHandler(log_stream_handler)
+
+
+def get_encoding_from_header(content_type):
+    content_type, params = cgi.parse_header(content_type)
+    if "charset" in params:
+        return params["charset"]
+
+
+_charset_flag = re.compile(r"""<meta.*?charset=["']*(.+?)["'>]""", flags=re.I)
+_pragma_flag = re.compile(r"""<meta.*?content=["']*;?charset=(.+?)["'>]""", flags=re.I)
+_xml_flag = re.compile(r"""^<\?xml.*?encoding=["']*(.+?)["'>]""")
+
+
+def get_encoding_from_content(content):
+    if isinstance(content, bytes):
+        content = content.decode("ascii", errors="ignore")
+    elif not isinstance(content, str):
+        raise ValueError("content should be bytes or str")
+    s = _charset_flag.search(content)
+    if s:
+        return s.group(1).strip()
+    s = _pragma_flag.search(content)
+    if s:
+        return s.group(1).strip()
+    s = _xml_flag.search(content)
+    if s:
+        return s.group(1).strip()
+
+
+def request_fingerprint(request):
+    sha1 = hashlib.sha1()
+    sha1.update(to_types(request.method))
+    sha1.update(to_types(str(request.url)))
+    sha1.update(request.body or b'')
+    return sha1.hexdigest()
+
+
+def to_types(data, encoding=None):
+    if isinstance(data, bytes):
+        return data
+    if isinstance(data, str):
+        return data.encode(encoding or "utf-8")
+    raise TypeError("Need bytes or str, got {}".format(type(data).__name__))
+
+
+def render_templatefile(path, **kwargs):
+    if path.endswith(".tmpl"):
+        with open(path, "rb") as f:
+            raw = f.read().decode("utf-8")
+
+        content = string.Template(raw).substitute(**kwargs)
+
+        render_path = path[:-len(".tmpl")]
+        with open(render_path, "wb") as f:
+            f.write(content.encode("utf-8"))
+        os.remove(path)
+
+
+_camelcase_invalid_chars = re.compile('[^a-zA-Z\d]')
+
+
+def string_camelcase(s):
+    return _camelcase_invalid_chars.sub('', s.title())
+
+
+def run_crawler(project_dir, **kwargs):
+    if not isfile(join(project_dir, "setup.cfg")):
+        raise FileNotFoundError("Cannot find 'setup.cfg' in {}".format(abspath(project_dir)))
+
+    config = Config(kwargs, priority="project")
+    configure_logging("xpaw", config)
+    cluster = LocalCluster(project_dir, config)
+    cluster.start()