+++ docs/changelog.rst
@@ -10,7 +10,7 @@ New features
 ~~~~~~~~~~~~
 
 - 使用继承Dupefilter的去重过滤器来实现去重功能，系统配置新增 ``dupefilter_cls`` 项，用于替换默认的去重过滤器
-- ``xpaw.utils.run`` 模块中新增 ``run_crawler`` 函数，便于在python代码中控制开启爬虫
+- ``xpaw.utils`` 模块中新增 ``run_crawler`` 函数，便于在python代码中控制开启爬虫
 
 Update
 ~~~~~~